+++ app/build.gradle
@@ -4,7 +4,6 @@ plugins {
     id 'androidx.navigation.safeargs.kotlin'
     id 'kotlin-kapt'
     id 'io.gitlab.arturbosch.detekt'
-    //id 'com.google.dagger.hilt.android'
 }
 
 android {
