+++ app/src/main/java/com/example/todoapp/ui/fragments/addEdit/viewmodel/AddEditUiState.kt
@@ -7,4 +7,7 @@ data class AddEditUiState(
     val text: String = "",
     val importance: Importance = Importance.COMMON,
     val deadline: Date? = null,
+    val isDeleteButtonEnabled: Boolean = false,
+    val isBottomSheetVisible: Boolean = false,
+    val isDatePickerDialogVisible: Boolean = false
 )
